Modification to topographic_plots.plot_head function for compatibility with channel_set.montage=None.
Modification to topographic_plots.plot_head and topographic_plots.plot_topography. Now both functions include an option to select if you want background on the figure or not. Default is False for compatibility with previous implementations where this option was not present.

Added option of bakground to mi_plots.plot_r2_topoplot